fix: list local folder in ContentBrowser
