+++ src/lib/components/Stores/workingFile.ts
@@ -1,0 +1,3 @@
+import { type Writable, writable } from 'svelte/store';
+
+export const workingFileStore: Writable<string> = writable('');
